requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.4
django-extensions==3.2.1
djangorestframework==3.14.0
drf-yasg==1.21.4
flake8==6.0.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
packaging==22.0
petl==1.7.12
psycopg2==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
pytz==2022.7
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.8.0
sqlparse==0.4.3
ujson==1.33
uritemplate==4.1.1
urllib3==1.26.13
